% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spongedown.R
\name{spongedown}
\alias{spongedown}
\title{SpongeDown}
\usage{
spongedown(text, htmlfile = tempfile(fileext = ".html"))
}
\arguments{
\item{text}{input ascii text}

\item{htmlfile}{output file to write html}
}
\description{
Render Markdown with support for svgbob ascii drawings.
}
\examples{
text <- readLines(system.file('examples/demo.bob', package = 'spongedown'))
html <- spongedown(text)
utils::browseURL(html)
}
\references{
Live demo: \url{http://ivanceras.github.io/svgbob-editor/}
}
